Support double for operator TopK + fix one bug in TopK implementation for GPU for double (#6220)
* Support double for operator TopK
* add static classes for topk/double
* fix cast issue in topk

+++ onnxruntime/core/providers/cuda/math/topk_impl.cu
@@ -214,7 +214,7 @@ __device__ int32_t Radix(const float* f, int64_t skip) {
 }
 
 __device__ int32_t Radix(const double* d, int64_t skip) {
-  return Radix((const double*)d, skip);
+  return Radix((const int64_t*)d, skip);
 }
 
 template<typename T>
